move DownloadWriter implementations to renter
